Moved crop functions to crop class
