+++ app/src/main/java/com/example/cropshot/MainActivity.java
@@ -105,13 +105,13 @@ public class MainActivity extends AppCompatActivity {
 
     public void onCropClick(View v) {
         try {
-            //Convert uri image to bitmap
-
+            // Create firebase object
+            FirebaseDetection firebaseDetectionObject = new FirebaseDetection(this);
 
             // Create a small version of the bitmap at the top of the screen
             // (Where the words instagram are) to scan for the text "instagram"
             Bitmap instaCrop = Bitmap.createBitmap(bitMap, 0, 0, bitMap.getWidth(), 200);
-            RunTextDetection(instaCrop);
+            firebaseDetectionObject.runTextDetection(instaCrop);
 
 
 
@@ -120,37 +120,23 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
-    private void cropIfImageDetected()
+    public void cropIfImageDetected()
     {
-        try {
-            // This function only gets called if firebase detects the string instagram in the image
+        System.out.println("cropIfImageDetected Called!");
 
-            System.out.println("Cropping");
+        Crop cropImg = new Crop();
+        croppedMap = cropImg.cropImage(contentURI, imageView, this);
 
-            // Call the FindBorder function for both top and bottom, to find the top and bottom border heights
-            int topCropInt = FindBorder(DIR.TOP);
-            int bottomCropInt = FindBorder(DIR.BOTTOM);
+        if(croppedMap == null)
+            return;
 
-
-            bottomCropInt = bitMap.getHeight() - bottomCropInt;
-
-            // Crop the top of the bitmap. Because bitmaps 0,0 starts in upper left, we must insert topCropInt as the
-            // Lower bounded value
-            croppedMap = Bitmap.createBitmap(bitMap, 0, topCropInt, bitMap.getWidth(), bitMap.getHeight() - topCropInt - bottomCropInt);
-
-            //saveImage(bitMap, "IMG300");
-            imageView.setImageBitmap(croppedMap);
-
-            Intent postcrop = new Intent(this, PostCropActivity.class);
-            ByteArrayOutputStream stream = new ByteArrayOutputStream();
-            croppedMap.compress(Bitmap.CompressFormat.JPEG, 100, stream);
-            byte[] bytes = stream.toByteArray();
-            postcrop.putExtra("cropBytes", bytes);
-            postcrop.putExtra("precropuri", contentURI.toString());
-            startActivity(postcrop);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        Intent postcrop = new Intent(this, PostCropActivity.class);
+        ByteArrayOutputStream stream = new ByteArrayOutputStream();
+        croppedMap.compress(Bitmap.CompressFormat.JPEG, 100, stream);
+        byte[] bytes = stream.toByteArray();
+        postcrop.putExtra("cropBytes", bytes);
+        postcrop.putExtra("precropuri", contentURI.toString());
+        startActivity(postcrop);
     }
 
     @Override
@@ -200,8 +186,6 @@ public class MainActivity extends AppCompatActivity {
 
     protected void saveImage(Bitmap finalBitmap) {
 
-
-
         String root = Environment.getExternalStorageDirectory().toString();
         File myDir = new File(root + "/saved_images");
         myDir.mkdirs();
@@ -224,93 +208,7 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
-    // Checks the color of the left and right pixel
-    // Returns true if the pixels are in a similar range
-    // And false otherwise
-    boolean CheckColor(int left, int right){
-        int leftSum = Color.red(left) + Color.green(left) + Color.blue(left);
-        int rightSum = Color.red(right) + Color.green(right) + Color.blue(right);
-        int diff = leftSum - rightSum;
 
-
-        if (diff < 10 && diff > -10) {
-            return true;
-        }
-        else
-            return false;
-       }
-
-    // Takes the bitmap, and given a direction (Top or bottom)
-    // Scans for pixels that are identical on the same line
-    // In order to find the top/bottom of an image
-    int FindBorder(DIR direction) {
-        // Get the middle position of the bitmap
-        int middleY = bitMap.getHeight() / 2;
-
-        if (direction == DIR.BOTTOM) {
-            // Let's start with I at the middle of the image, and move positively until we reach the bottom
-            for (int i = middleY; i < bitMap.getHeight(); i++)
-            {
-
-                // Generate the single rowed bitmap
-                Bitmap subMap = Bitmap.createBitmap(bitMap, 0, i, bitMap.getWidth(), 1);
-                int pixel = subMap.getPixel(0,0);
-
-                if(SolidRow(subMap) && Color.red(pixel) == Color.blue(pixel) && Color.red(pixel) == Color.green(pixel))
-                {
-                    return i;
-                }
-            }
-
-            //DIR == TOP
-        } else {
-            // Let's start with I at the middle of the image, and move negatively until we reach the top
-            for (int i = middleY; i > 0; i--)
-            {
-
-                // Generate the single rowed bitmap
-                Bitmap subMap = Bitmap.createBitmap(bitMap, 0, i, bitMap.getWidth(), 1);
-                int pixel = subMap.getPixel(0,0);
-
-                // if the colors are the same we want to return our i value for the top
-                // Also if the color is white (255,255,255), as that designates top of image.
-                if(SolidRow(subMap))
-                {
-                    return i;
-                }
-            }
-        }
-        return 0;
-    }
-
-    // NOTE: This doesn't actually check if all the pixels are the same!
-    // It only checks if the pixels across from eachother are, i.e.
-    // (left, right), (left + 1, right - 1)
-
-    // Given a row of pixels in a bitmap, return true if all pixels
-    // Are the same color, otherwise return false.
-    boolean SolidRow(Bitmap row)
-    {
-        //Getting length and height for the bitmap
-        int length = row.getWidth();
-        int height = row.getHeight();
-        int max = length - 1;
-        int pixel = row.getPixel(0,0);
-
-        //Iterates through the bitmap row
-        for(int i = 1; i < (length - 1); i++) {
-            //Gets variables
-            int left_pixel = row.getPixel(0, height - 1);
-            int right_pixel = row.getPixel(i, height - 1);
-
-
-            //Checks if the pixels are the same color or if the pixels meet
-            if (!CheckColor(left_pixel, right_pixel)) {
-                return false;
-            }
-        }
-      return true;
-    }
 
 
     public Bitmap cropSolidRow(Bitmap bitmap, int cropHeight, DIR direction){
@@ -333,62 +231,6 @@ public class MainActivity extends AppCompatActivity {
 
             Bitmap newBitmap = Bitmap.createBitmap(bitmap, 0, 0, bitmapWidth, numOfRows);
             return newBitmap;
-        }
-    }
-
-    private void RunTextDetection(Bitmap croppedMap)
-    {
-        FirebaseVisionImage image = FirebaseVisionImage.fromBitmap(croppedMap);
-        FirebaseVisionTextRecognizer detector = FirebaseVision.getInstance().getOnDeviceTextRecognizer();
-
-        detector.processImage(image).addOnSuccessListener(
-                new OnSuccessListener<FirebaseVisionText>() {
-                    @Override
-                    public void onSuccess(FirebaseVisionText texts) {
-                        System.out.println("Succeeded process image!");
-                        ProcessTextRecognitionResults(texts);
-
-                    }
-                }
-
-        ).addOnFailureListener(
-                new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        System.out.println("Failed to parse image");
-                    }
-                }
-        );
-    }
-
-    private void ProcessTextRecognitionResults(FirebaseVisionText texts) {
-        System.out.println(texts.getText());
-
-        // Get all of the blocks in the current text
-        List<FirebaseVisionText.TextBlock> blocks = texts.getTextBlocks();
-
-        if (blocks.size() == 0)
-            return;
-
-        for (int i = 0; i < blocks.size(); i++) {
-
-            List<FirebaseVisionText.Line> lines = blocks.get(i).getLines();
-            for (int j = 0; j < lines.size(); j++) {
-                List<FirebaseVisionText.Element> elements = lines.get(j).getElements();
-                for (int k = 0; k < elements.size(); k++) {
-                    // The text detection is not perfect, especially in the case of an image that
-                    // is slightly more blurry. When weird issues come up, we add a case to account
-                    // for those, hence the "instagam"
-                    String str = elements.get(k).getText().toLowerCase();
-                    if (str.contains("instagram") ||
-                            str.contains("instagam") ||
-                            str.contains("nstagam")) {
-                        System.out.println("Instagram image detected!");
-                        cropIfImageDetected();
-                        return;
-                    }
-                }
-            }
         }
     }
