+++ app/src/main/java/com/example/cropshot/Save.java
@@ -1,4 +1,19 @@
 package com.example.cropshot;
 
 public class Save {
+
+    public void saveAsNew()
+    {
+
+    }
+
+    public void Overwrite()
+    {
+
+    }
+
+    public void SaveAs(String name)
+    {
+
+    }
 }
